+++ c++/libpente/board.h
@@ -45,7 +45,7 @@ public:
 	const VecVec &getVectors() const;
 	const Vector* winner(uint32_t &nplayer) const;
 	uint32_t size() const {
-		return rep.size();
+		return (uint32_t)rep.size();
 	}
 	void getCaptures(uint32_t x, uint32_t y, CaptureVec &captures);
